# code/logistics/public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"expectedRouteJSON": {
				"properties": {
					"extraChargingTime": {
						"type": "number",
						"format": "double"
					},
					"energyConsumed": {
						"type": "number",
						"format": "double"
					},
					"timeRequired": {
						"type": "number",
						"format": "double"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"idEnd": {
						"type": "string"
					},
					"idStart": {
						"type": "string"
					},
					"idRoute": {
						"type": "string"
					}
				},
				"required": [
					"extraChargingTime",
					"energyConsumed",
					"timeRequired",
					"distance",
					"idEnd",
					"idStart",
					"idRoute"
				],
				"type": "object"
			},
			"expectedBodyRoute": {
				"properties": {
					"extraChargingTime": {
						"type": "number",
						"format": "double"
					},
					"energyConsumed": {
						"type": "number",
						"format": "double"
					},
					"timeRequired": {
						"type": "number",
						"format": "double"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"idEnd": {
						"type": "string"
					},
					"idStart": {
						"type": "string"
					},
					"idRoute": {
						"type": "string"
					}
				},
				"required": [
					"extraChargingTime",
					"energyConsumed",
					"timeRequired",
					"distance",
					"idEnd",
					"idStart",
					"idRoute"
				],
				"type": "object"
			},
			"expectedPatchBodyRoute": {
				"properties": {
					"extraChargingTime": {
						"type": "number",
						"format": "double"
					},
					"energyConsumed": {
						"type": "number",
						"format": "double"
					},
					"timeRequired": {
						"type": "number",
						"format": "double"
					},
					"distance": {
						"type": "number",
						"format": "double"
					},
					"idEnd": {
						"type": "string"
					},
					"idStart": {
						"type": "string"
					}
				},
				"required": [
					"extraChargingTime",
					"energyConsumed",
					"timeRequired",
					"distance",
					"idEnd",
					"idStart"
				],
				"type": "object"
			},
			"SliceRouteType": {
				"properties": {
					"idEnd": {
						"type": "string"
					},
					"idStart": {
						"type": "string"
					}
				},
				"required": [
					"idEnd",
					"idStart"
				],
				"type": "object"
			},
			"expectedTripJSON": {
				"properties": {
					"deliveries": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"routes": {
						"items": {
							"$ref": "#/components/schemas/SliceRouteType"
						},
						"type": "array"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"registration": {
						"type": "string"
					},
					"idTrip": {
						"type": "string"
					}
				},
				"required": [
					"deliveries",
					"routes",
					"date",
					"registration",
					"idTrip"
				],
				"type": "object"
			},
			"expectedBodyTrip": {
				"properties": {
					"deliveries": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"routes": {
						"items": {
							"$ref": "#/components/schemas/SliceRouteType"
						},
						"type": "array"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"registration": {
						"type": "string"
					},
					"idTrip": {
						"type": "string"
					}
				},
				"required": [
					"deliveries",
					"routes",
					"date",
					"registration",
					"idTrip"
				],
				"type": "object"
			},
			"expectedTruckJSON": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"autonomy": {
						"type": "number",
						"format": "double"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"tare": {
						"type": "number",
						"format": "double"
					},
					"registration": {
						"type": "string"
					}
				},
				"required": [
					"active",
					"autonomy",
					"capacity",
					"tare",
					"registration"
				],
				"type": "object"
			},
			"expectedBodyTruck": {
				"properties": {
					"active": {
						"type": "boolean"
					},
					"autonomy": {
						"type": "number",
						"format": "double"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"tare": {
						"type": "number",
						"format": "double"
					},
					"registration": {
						"type": "string"
					}
				},
				"required": [
					"active",
					"autonomy",
					"capacity",
					"tare",
					"registration"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "logistics",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/routes": {
			"get": {
				"operationId": "GetRoutes",
				"responses": {
					"200": {
						"description": "JSON with all the routes",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/expectedRouteJSON"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Returns all the routes that exist",
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateRoute",
				"responses": {
					"200": {
						"description": "Returns a JSON with the created route",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedBodyRoute"
								}
							}
						}
					}
				},
				"summary": "Creates a new route from the given body and returns the created route",
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedBodyRoute"
							}
						}
					}
				}
			}
		},
		"/routes/pag/ination": {
			"get": {
				"operationId": "GetRoutesByPagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalRecords": {
											"type": "number",
											"format": "double"
										},
										"routesList": {
											"items": {
												"$ref": "#/components/schemas/expectedRouteJSON"
											},
											"type": "array"
										}
									},
									"required": [
										"totalRecords",
										"routesList"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageRecords",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/routes/{idRoute}": {
			"get": {
				"operationId": "GetRouteById",
				"responses": {
					"200": {
						"description": "Returns a JSON with the route information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedRouteJSON"
								}
							}
						}
					}
				},
				"summary": "This method searches for a route from the given id",
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": [
					{
						"description": "idRoute of the route",
						"in": "path",
						"name": "idRoute",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateRouteById",
				"responses": {
					"200": {
						"description": "Updated route",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedRouteJSON"
								}
							}
						}
					}
				},
				"summary": "Searches for the route with the given idRoute and updates it with any of the parameters given in the body",
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id of the route to be updated",
						"in": "path",
						"name": "idRoute",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Can update any of the parameters of the route, not all of them are required",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedPatchBodyRoute",
								"description": "Can update any of the parameters of the route, not all of them are required"
							}
						}
					}
				}
			}
		},
		"/trips": {
			"get": {
				"operationId": "GetTrips",
				"responses": {
					"200": {
						"description": "JSON with all the trips",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/expectedTripJSON"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Returns all the trips that exist",
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTrip",
				"responses": {
					"200": {
						"description": "Returns a JSON with the created trip",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedBodyTrip"
								}
							}
						}
					}
				},
				"summary": "Creates a new trip from the given body and returns the created trip",
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedBodyTrip"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateTripById",
				"responses": {
					"200": {
						"description": "Returns a JSON with the updated trip",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedBodyTrip"
								}
							}
						}
					}
				},
				"summary": "Will update a trip with the given body else if there is no trip with the given idTrip it will create a new trip",
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "body of the trip",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedBodyTrip",
								"description": "body of the trip"
							}
						}
					}
				}
			}
		},
		"/trips/records/pagination/data": {
			"get": {
				"operationId": "GetTripsByPagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalRecords": {
											"type": "number",
											"format": "double"
										},
										"tripList": {
											"items": {
												"$ref": "#/components/schemas/expectedTripJSON"
											},
											"type": "array"
										}
									},
									"required": [
										"totalRecords",
										"tripList"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageRecords",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trips/{idTrip}": {
			"get": {
				"operationId": "GetTripById",
				"responses": {
					"200": {
						"description": "Returns a JSON with the trip information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedTripJSON"
								}
							}
						}
					}
				},
				"summary": "This method searches for a trip from the given id",
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [
					{
						"description": "idTrip of the trip",
						"in": "path",
						"name": "idTrip",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/trips/{registration}/{date}": {
			"get": {
				"operationId": "GetTripByRegDate",
				"responses": {
					"200": {
						"description": "Returns a JSON with the trip information",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedTripJSON"
								}
							}
						}
					}
				},
				"summary": "This method searches for a trip from the given registration and date",
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [
					{
						"description": "registration of the truck",
						"in": "path",
						"name": "registration",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/trips/optimize/{registration}/{date}": {
			"put": {
				"operationId": "OptimizeTrip",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedBodyTrip"
								}
							}
						}
					}
				},
				"tags": [
					"Trips"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "registration",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/trucks": {
			"get": {
				"operationId": "GetTrucks",
				"responses": {
					"200": {
						"description": "JSON with all the trucks",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/expectedTruckJSON"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Returns all the trucks that exist",
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateTruck",
				"responses": {
					"200": {
						"description": "Returns a JSON with the created truck",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedBodyTruck"
								}
							}
						}
					}
				},
				"summary": "Creates a truck from the given body and returns the created truck",
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedBodyTruck"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateTruckByRegistration",
				"responses": {
					"200": {
						"description": "Returns a JSON with the updated truck",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedTruckJSON"
								}
							}
						}
					}
				},
				"summary": "Will update a truck with the given body else if there is no truck with the given registration it will create a new truck",
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "- Truck body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/expectedBodyTruck",
								"description": "- Truck body"
							}
						}
					}
				}
			}
		},
		"/trucks/pag/ination": {
			"get": {
				"operationId": "GetTrucksByPagination",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"totalRecords": {
											"type": "number",
											"format": "double"
										},
										"trucksList": {
											"items": {
												"$ref": "#/components/schemas/expectedTruckJSON"
											},
											"type": "array"
										}
									},
									"required": [
										"totalRecords",
										"trucksList"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageRecords",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trucks/{registration}": {
			"get": {
				"operationId": "GetTruckByRegistration",
				"responses": {
					"200": {
						"description": "Returns a JSON with the truck if its found else returns an error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedTruckJSON"
								}
							}
						}
					}
				},
				"summary": "Gets a truck by its registration",
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The registration of the truck",
						"in": "path",
						"name": "registration",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/trucks/status/{registration}": {
			"patch": {
				"operationId": "ChangeActiveStatus",
				"responses": {
					"200": {
						"description": "Returns a JSON with the updated truck",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/expectedTruckJSON"
								}
							}
						}
					}
				},
				"summary": "Changes the active status of a truck",
				"tags": [
					"Trucks"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The registration of the truck",
						"in": "path",
						"name": "registration",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}